+++ packages/expo-ui/src/swift-ui/ContextMenu/utils.ts
@@ -89,10 +89,12 @@ function createButtonElement(
   if (props.onPress) {
     handlers[uuid] = { onPress: props.onPress };
   }
-
+  if (typeof props.children !== 'string') {
+    throw new Error('ContextMenu Button only supports string children');
+  }
   return {
     contextMenuElementID: uuid,
-    button: transformButtonProps(props),
+    button: transformButtonProps(props, props.children),
   };
 }
 
@@ -135,10 +137,13 @@ function createSubmenuElement(
   props: SubmenuProps,
   handlers: EventHandlers
 ): MenuElement {
+  if (typeof props.button.props.children !== 'string') {
+    throw new Error('ContextMenu Submenu Button only supports string children');
+  }
   return {
     contextMenuElementID: uuid,
     submenu: {
-      button: transformButtonProps(props.button.props),
+      button: transformButtonProps(props.button.props, props.button.props.children),
       elements: transformChildrenToElementArray(props.children, handlers),
     },
   };